Significant refactor for abstract HpcJobManager class and separate HPC module for PBS and SLURM utilities separate from the subprocess manager and local parallel execution utilities.
