+++ rex/__init__.py
@@ -18,7 +18,7 @@ from rex.resource_extraction import (ResourceX, MultiYearResourceX,
                                      NSRDBX, MultiFileNSRDBX, MultiYearNSRDBX,
                                      WindX, MultiFileWindX, MultiYearWindX,
                                      WaveX, MultiYearWaveX, TemporalStats)
-from rex.utilities import init_logger, init_mult, SpawnProcessPool, SLURM
+from rex.utilities import init_logger, init_mult, SpawnProcessPool, SLURM, PBS
 
 from rex.version import __version__
