+++ rex/utilities/__init__.py
@@ -2,7 +2,8 @@
 """
 rex utilities.
 """
-from .execution import SpawnProcessPool, SLURM
+from .execution import SpawnProcessPool
+from .hpc import SLURM, PBS
 from .loggers import init_logger, init_mult, setup_logger, log_mem
 from .solar_position import SolarPosition
 from .toml_parser import TOMLParser
